# docs/api-documentation.yaml
openapi: 3.0.0
info:
  title: Training Management System API
  description: Complete REST API for managing corporate training programs with trainers, projects, engagements, and attendance tracking
  version: 1.0.0
  contact:
    name: API Support
    email: support@trainermanagement.com
servers:
  - url: http://localhost:8000/v1
    description: Local development server
  - url: https://api.trainermanagement.com/v1
    description: Production server
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Trainer:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        expertise:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [available, partially_allocated, fully_allocated, on_leave]
        max_batches:
          type: integer
        current_batches:
          type: integer
        employment_type:
          type: string
          enum: [full-time, part-time, freelance, intern]
        experience_level:
          type: string
          enum: [junior, mid, senior]
        gender:
          type: string
          enum: [male, female]
        join_date:
          type: string
          format: date
        avatar:
          type: string
...
security:
  - BearerAuth: []
paths:
...
  /trainers/:
    get:
      tags:
        - Trainers
      summary: Get all trainers
      parameters:
        - in: query
          name: skip
          schema:
            type: integer
            default: 0
        - in: query
          name: limit
          schema:
            type: integer
            default: 100
        - in: query
          name: status
          schema:
            type: string
        - in: query
          name: expertise
          schema:
            type: string
      responses:
        '200':
          description: List of trainers
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Trainer'
                  metadata:
                    type: object
                    properties:
                      total:
                        type: integer
                      page:
                        type: integer
                      limit:
                        type: integer

    post:
      tags:
        - Trainers
      summary: Create new trainer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                email:
                  type: string
                phone:
                  type: string
                expertise:
                  type: array
                  items:
                    type: string
                employment_type:
                  type: string
                experience_level:
                  type: string
                max_batches:
                  type: integer
                gender:
                  type: string
                join_date:
                  type: string
                  format: date
      responses:
        '201':
          description: Trainer created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Trainer'
  /trainers/{trainer_id}:
    get:
      tags:
        - Trainers
      summary: Get trainer by ID
      parameters:
        - in: path
          name: trainer_id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Trainer details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Trainer'
        '404':
          description: Trainer not found

    put:
      tags:
        - Trainers
      summary: Update trainer
      parameters:
        - in: path
          name: trainer_id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Trainer updated

    delete:
      tags:
        - Trainers
      summary: Delete trainer
      parameters:
        - in: path
          name: trainer_id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Trainer deleted
...